requirements.txt
Django==5.0.7
django-js-asset==2.2.0
dnspython==2.6.1
docutils==0.21.2
markdown2==2.5.0
Pillow==10.4.0
py3-validate-email
requests==2.32.3
django-appointment==3.6.3
django-payment-manager==1.0.1
python-dotenv==1.0.1
django-simple-history==3.7.0
xhtml2pdf==0.2.16
pypdf2~=3.0.1
endesive==2.17.2
cryptography==42.0.7
django-simple-captcha==0.6.0
